user grid list & front-end permission table check

+++ TrinityService/src/main/java/com/netpro/trinity/service/member/controller/TrinityuserController.java
@@ -1,6 +1,5 @@
 package com.netpro.trinity.service.member.controller;
 
-import java.lang.reflect.InvocationTargetException;
 import java.sql.SQLException;
 
 import javax.servlet.http.HttpServletRequest;
@@ -20,6 +19,7 @@ import com.netpro.ac.ACException;
 import com.netpro.trinity.service.dto.FilterInfo;
 import com.netpro.trinity.service.member.entity.Trinityuser;
 import com.netpro.trinity.service.member.service.TrinityuserService;
+import com.netpro.trinity.service.permission.feign.PermissionClient;
 
 @RestController  //宣告一個Restful Web Service的Resource
 @RequestMapping("/trinity-user")
@@ -28,6 +28,9 @@ public class TrinityuserController {
 		
 	@Autowired
 	private TrinityuserService service;
+	
+	@Autowired
+	private PermissionClient permissionClient;
 	
 	@GetMapping("/findAll")
 	public ResponseEntity<?> findAllUsers() {
@@ -52,7 +55,11 @@ public class TrinityuserController {
 	@GetMapping("/findByUid")
 	public ResponseEntity<?> findUserByUid(String uid) {
 		try {
-			return ResponseEntity.ok(this.service.getByUid(uid));
+			if(this.permissionClient.isRootOrAdminOrAccountManager()) {
+				return ResponseEntity.ok(this.service.getByUid(uid));
+			}else {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You do not have 'Root, Admin or AccountManager' Permission!");
+			}
 		}catch(IllegalArgumentException e) {
 			TrinityuserController.LOGGER.error("IllegalArgumentException; reason was:", e);
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
@@ -61,14 +68,11 @@ public class TrinityuserController {
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
 		}
 	}
-  
-	@GetMapping("/findByName")
-	public ResponseEntity<?> findUsersByName(String name) {
+	
+	@PostMapping("/findByFilter")
+	public ResponseEntity<?> findUsersByFilter(@RequestBody FilterInfo filter) {
 		try {
-			return ResponseEntity.ok(this.service.getByName(name));
-		}catch(IllegalArgumentException e) {
-			TrinityuserController.LOGGER.error("IllegalArgumentException; reason was:", e);
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
+			return this.service.getByFilter(filter);
 		}catch(Exception e) {
 			TrinityuserController.LOGGER.error("Exception; reason was:", e);
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
@@ -88,49 +92,14 @@ public class TrinityuserController {
 		}
 	}
 	
-	@GetMapping("/findByType")
-	public ResponseEntity<?> findUsersByType(String type) {
-		try {
-			return ResponseEntity.ok(this.service.getByUserType(type));
-		}catch(IllegalArgumentException e) {
-			TrinityuserController.LOGGER.error("IllegalArgumentException; reason was:", e);
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
-		}catch(Exception e) {
-			TrinityuserController.LOGGER.error("Exception; reason was:", e);
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
-		}
-	}
-	
-	@PostMapping("/findByFilter")
-	public ResponseEntity<?> findUsersByFilter(@RequestBody FilterInfo filter) {
-		try {
-			return this.service.getByFilter(filter);
-		}catch(SecurityException e) {
-			TrinityuserController.LOGGER.error("SecurityException; reason was:", e);
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
-		}catch(NoSuchMethodException e) {
-			TrinityuserController.LOGGER.error("NoSuchMethodException; reason was:", e);
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
-		}catch(IllegalAccessException e) {
-			TrinityuserController.LOGGER.error("IllegalAccessException; reason was:", e);
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
-		}catch(InvocationTargetException e) {
-			TrinityuserController.LOGGER.error("InvocationTargetException; reason was:", e);
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
-		}catch(IllegalArgumentException e) {
-			TrinityuserController.LOGGER.error("IllegalArgumentException; reason was:", e);
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
-		}catch(Exception e) {
-			TrinityuserController.LOGGER.error("Exception; reason was:", e);
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
-		}
-	}
-	
 	@PostMapping("/add")
 	public ResponseEntity<?> addRole(HttpServletRequest request, @RequestBody Trinityuser user) {
 		try {
-			user = this.service.add(request, user);
-			return ResponseEntity.ok(user);
+			if(this.permissionClient.isRootOrAdminOrAccountManager()) {
+				return ResponseEntity.ok(this.service.add(request, user));
+			}else {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You do not have 'Root, Admin or AccountManager' Permission!");
+			}
 		}catch(ACException e) {
 			TrinityuserController.LOGGER.error("ACException; reason was:", e);
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
@@ -149,7 +118,25 @@ public class TrinityuserController {
 	@PostMapping("/edit")
 	public ResponseEntity<?> editRole(HttpServletRequest request, @RequestBody Trinityuser user) {
 		try {
-			return ResponseEntity.ok(this.service.edit(request, user));
+			String useruid = user.getUseruid();
+			Trinityuser old_user = this.service.getByUid(useruid);
+			Trinityuser reqUser = this.service.getUserFormRequest(request);
+			
+			if(this.permissionClient.isAdminEditMode()) {
+				if(this.permissionClient.isRootOrAdminOrAccountManager()) {
+					return ResponseEntity.ok(this.service.edit(request, user));
+				}else {
+					return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You do not have 'Root, Admin or AccountManager' Permission!");
+				}
+			}else {
+				if(this.permissionClient.isRootOrAccountManager() || 
+						old_user.getCreateduseruid().trim().equals(reqUser.getUseruid().trim()) || 
+						useruid.trim().equals(reqUser.getUseruid().trim())) {
+					return ResponseEntity.ok(this.service.edit(request, user));
+				}else {
+					return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("This account is created by root or another admin!");
+				}
+			}
 		}catch(ACException e) {
 			TrinityuserController.LOGGER.error("ACException; reason was:", e);
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
@@ -166,9 +153,18 @@ public class TrinityuserController {
 	}
 	
 	@GetMapping("/delete")
-	public ResponseEntity<?> deleteUserByUid(String uid) {
+	public ResponseEntity<?> deleteUserByUid(HttpServletRequest request, String uid) {
 		try {
-			this.service.deleteByUid(uid);
+			Trinityuser old_user = this.service.getByUid(uid);
+			Trinityuser reqUser = this.service.getUserFormRequest(request);
+			
+			if(this.permissionClient.isRootOrAccountManager() || 
+					old_user.getCreateduseruid().trim().equals(reqUser.getUseruid().trim()) || 
+					uid.trim().equals(reqUser.getUseruid().trim())) {
+				this.service.deleteByUid(uid);
+			}else {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("This account is created by root or another admin!");
+			}
 		}catch(IllegalArgumentException e) {
 			TrinityuserController.LOGGER.error("IllegalArgumentException; reason was:", e);
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
@@ -182,7 +178,16 @@ public class TrinityuserController {
 	@GetMapping("/lock")
 	public ResponseEntity<?> lockUserByID(HttpServletRequest request, String userid, Boolean lock) {
 		try {
-			return ResponseEntity.ok(this.service.lockByUserID(request, userid, lock));
+			Trinityuser old_user = this.service.getByID(userid);
+			Trinityuser reqUser = this.service.getUserFormRequest(request);
+			
+			if(this.permissionClient.isRootOrAccountManager() || 
+					old_user.getCreateduseruid().trim().equals(reqUser.getUseruid().trim()) || 
+					old_user.getUseruid().trim().equals(reqUser.getUseruid().trim())) {
+				return ResponseEntity.ok(this.service.lockByUserID(request, userid, lock));
+			}else {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("This account is created by root or another admin!");
+			}
 		}catch(IllegalArgumentException e) {
 			TrinityuserController.LOGGER.error("IllegalArgumentException; reason was:", e);
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
